Numerous fixes. Added new .data to old .data pointer scanning and modification
Nginx works at functions:
    - ngx_http_process_request_line
    - ngx_http_handler
    - ngx_single_process_cycle

Issues: #42, #45, #46, #9

Initially I thought it was .bss pointers but it's actually .data pointers. Referenced in issue
42.

Numerous fixes. Added support for accept4 argument emulation as we needed the ip address to be passed out
via the argument. We were not simulating that correctly.
Also fixed another bug where we were running out of file descriptors after 1000+ rounds.
This occurred because when cloning we were cloning file descriptors as well, which was not needed.
Fixed an fopen without fclose pair.