Added Timing capabilities
Issues: #50

We wanted to time which part of lmvx_start was the main bottleneck. By default this timing capability is turned off.
For a single invocation on mvx the timing works, but when we attempt to invoke over multiple runs the execution diverges.
For now at least we have the numbers for a single run. Since this is only a timing instrumentation, we defer the fixing of this
to if and when we need to time the overheads of lmvx_start over multiple runs. Enable by setting the flag _TIME_LMVX to 1 in config.h.