Added PLT patching and Protection
Issue: #51

We add functions to read out the gotplt addresses and then patch them as immediate jump instructions in the plt.
This enables us to hide where libc and the monitor trampoline are located. The plt slots are patched with a movabs
of gotplt entries to $rax followed by a jumpq to rax.