Added additional PLT patching to the workshop paper subdir
We patch the PLT to send all libc calls to a trampoline defined in
trampoline.s. This trampoline provides a universal way tor us to
switch MPK callgates. The one downside right now is it uses a callq
instruction which means that any libc calls that use more than 6 arguments
will have a presumed [%rbp+x] location to get the data but that presumption
will be incorrect since the %rip will be pushed onto the stack due to callq.