Fixed issue where push of slot_number greater than 0x7f rolled over
Issues: #55

Another fix related to the unified trampoline. The push of the slot_number was
pushing a negative number when greater than 0x7f as we were using the wrong opcode,
0x6A instead of 0x68. However, if we switch to 0x68 we need more bytes than we have space.

We have split the plt stub into two parts. The first part is the individual libc call stub,
in charge of pushing rbx, rax, and then jumping to what used to be the ldso resolver to resolve
symbols via lazy binding. Musl doesn't use this feature so we can overwrite its entry in the plt
to perform the actual jump to the trampoline code.

Refer to the writeup in github for more details.