Added multithreading test/benchmark
Issue: #57

Benchmark added to verify that callgate using TLS stack works.
We use libthreadtest.so to provide one plt entry, library_func which is
gated with the trampoline. This library function reads and prints the
address of a stack variable, allowing us to verify that the stacks are in
different locations. To ensure there is no cross-write between the safestacks,
we also write in a loop to a variable and print out the final result. The library
function itself is called by 2 threads spawned in test/thread_test.c.